Change interface, that Validators implement to their respective Constraint interface
Some Validators implemented the interface the TransportConstraint, which seems to be written to be implemented only by TransportValidator. It shows up in the other Validators presumably because copy/paste error?
It didn't seem to harm the function of the lib but newer hibernate validator versions don't allow this anymore so we have to fix it.

prepares update for #16